+++ rocketmq-controller/src/processor/broker_processor.rs
@@ -15,6 +15,7 @@
 use std::sync::Arc;
 use std::time::SystemTime;
 
+use rocketmq_rust::ArcMut;
 use tracing::debug;
 use tracing::error;
 use tracing::info;
@@ -41,12 +42,12 @@ pub struct RegisterBrokerProcessor {
     metadata: Arc<MetadataStore>,
 
     /// Raft controller
-    raft: Arc<RaftController>,
+    raft: ArcMut<RaftController>,
 }
 
 impl RegisterBrokerProcessor {
     /// Create a new register broker processor
-    pub fn new(metadata: Arc<MetadataStore>, raft: Arc<RaftController>) -> Self {
+    pub fn new(metadata: Arc<MetadataStore>, raft: ArcMut<RaftController>) -> Self {
         Self { metadata, raft }
     }
 
@@ -119,12 +120,12 @@ pub struct UnregisterBrokerProcessor {
     metadata: Arc<MetadataStore>,
 
     /// Raft controller
-    raft: Arc<RaftController>,
+    raft: ArcMut<RaftController>,
 }
 
 impl UnregisterBrokerProcessor {
     /// Create a new unregister broker processor
-    pub fn new(metadata: Arc<MetadataStore>, raft: Arc<RaftController>) -> Self {
+    pub fn new(metadata: Arc<MetadataStore>, raft: ArcMut<RaftController>) -> Self {
         Self { metadata, raft }
     }
 
@@ -226,12 +227,12 @@ pub struct ElectMasterProcessor {
     metadata: Arc<MetadataStore>,
 
     /// Raft controller
-    raft: Arc<RaftController>,
+    raft: ArcMut<RaftController>,
 }
 
 impl ElectMasterProcessor {
     /// Create a new elect master processor
-    pub fn new(metadata: Arc<MetadataStore>, raft: Arc<RaftController>) -> Self {
+    pub fn new(metadata: Arc<MetadataStore>, raft: ArcMut<RaftController>) -> Self {
         Self { metadata, raft }
     }