+++ rocketmq-controller/src/processor/topic_processor.rs
@@ -14,6 +14,7 @@
 
 use std::sync::Arc;
 
+use rocketmq_rust::ArcMut;
 use tracing::error;
 use tracing::info;
 
@@ -37,12 +38,12 @@ pub struct CreateTopicProcessor {
     metadata: Arc<MetadataStore>,
 
     /// Raft controller
-    raft: Arc<RaftController>,
+    raft: ArcMut<RaftController>,
 }
 
 impl CreateTopicProcessor {
     /// Create a new create topic processor
-    pub fn new(metadata: Arc<MetadataStore>, raft: Arc<RaftController>) -> Self {
+    pub fn new(metadata: Arc<MetadataStore>, raft: ArcMut<RaftController>) -> Self {
         Self { metadata, raft }
     }
 
@@ -117,12 +118,12 @@ pub struct UpdateTopicProcessor {
     metadata: Arc<MetadataStore>,
 
     /// Raft controller
-    raft: Arc<RaftController>,
+    raft: ArcMut<RaftController>,
 }
 
 impl UpdateTopicProcessor {
     /// Create a new update topic processor
-    pub fn new(metadata: Arc<MetadataStore>, raft: Arc<RaftController>) -> Self {
+    pub fn new(metadata: Arc<MetadataStore>, raft: ArcMut<RaftController>) -> Self {
         Self { metadata, raft }
     }
 
@@ -188,12 +189,12 @@ pub struct DeleteTopicProcessor {
     metadata: Arc<MetadataStore>,
 
     /// Raft controller
-    raft: Arc<RaftController>,
+    raft: ArcMut<RaftController>,
 }
 
 impl DeleteTopicProcessor {
     /// Create a new delete topic processor
-    pub fn new(metadata: Arc<MetadataStore>, raft: Arc<RaftController>) -> Self {
+    pub fn new(metadata: Arc<MetadataStore>, raft: ArcMut<RaftController>) -> Self {
         Self { metadata, raft }
     }